clean up Link component structure
