+++ src/core/Link/Link/Link.baseStyles.ts
@@ -1,0 +1,6 @@
+import { css } from 'styled-components';
+import { baseStyles } from '../BaseLink/BaseLink.baseStyles';
+
+export const linkExternalStyles = css`
+  ${baseStyles}
+`;
