+++ src/core/Link/LinkExternal/LinkExternal.test.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { render } from '@testing-library/react';
-import { axeTest } from '../../utils/test/axe';
+import { axeTest } from '../../../utils/test/axe';
 
 import { LinkExternal } from './LinkExternal';
