+++ src/core/Link/index.ts
@@ -1,0 +1,3 @@
+export { Link, LinkProps } from './Link/Link';
+export { LinkExternal, LinkExternalProps } from './LinkExternal/LinkExternal';
+export { LinkSkip, LinkSkipProps } from './LinkSkip/LinkSkip';
